# server/swagger.yaml
openapi: 3.0.0
info:
  title: Smoking Support System API
  description: API documentation for the Smoking Support System platform
  version: 1.0.0
  contact:
    email: support@smokingsupport.com
servers:
  - url: http://localhost:5000/api
    description: Development server
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        coachId:
          type: integer
        scheduledTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        note:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        progressId:
          type: integer
        content:
          type: string
        sentAt:
          type: string
          format: date-time
        isRead:
          type: boolean
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /bookings:
    post:
      tags:
        - Bookings
      summary: Create booking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
    get:
      tags:
        - Bookings
      summary: Get user's bookings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
...
  /messages:
    post:
      tags:
        - Messages
      summary: Send message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    get:
      tags:
        - Messages
      summary: Get user's messages
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
...